remarcable/requirements.txt
asgiref==3.7.2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
install==1.3.5
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
